+++ lib/ts/recipe/thirdpartyemailpassword/types.ts
@@ -33,6 +33,7 @@ import {
     TypeInput as EmailDeliveryTypeInput,
     TypeInputWithService as EmailDeliveryTypeInputWithService,
 } from "../../ingredients/emaildelivery/types";
+import { GeneralErrorResponse } from "../../types";
 
 export type User = {
     id: string;
@@ -147,13 +148,7 @@ export type RecipeInterface = {
             isVerified: boolean;
         };
         userContext: any;
-    }): Promise<
-        | { status: "OK"; createdNewUser: boolean; user: User }
-        | {
-              status: "FIELD_ERROR";
-              error: string;
-          }
-    >;
+    }): Promise<{ status: "OK"; createdNewUser: boolean; user: User }>;
 
     emailPasswordSignUp(input: {
         email: string;
@@ -206,10 +201,13 @@ export type APIInterface = {
               provider: TypeProvider;
               options: ThirdPartyAPIOptions;
               userContext: any;
-          }) => Promise<{
-              status: "OK";
-              url: string;
-          }>);
+          }) => Promise<
+              | {
+                    status: "OK";
+                    url: string;
+                }
+              | GeneralErrorResponse
+          >);
 
     emailPasswordEmailExistsGET:
         | undefined
@@ -217,10 +215,13 @@ export type APIInterface = {
               email: string;
               options: EmailPasswordAPIOptions;
               userContext: any;
-          }) => Promise<{
-              status: "OK";
-              exists: boolean;
-          }>);
+          }) => Promise<
+              | {
+                    status: "OK";
+                    exists: boolean;
+                }
+              | GeneralErrorResponse
+          >);
 
     generatePasswordResetTokenPOST:
         | undefined
@@ -231,9 +232,12 @@ export type APIInterface = {
               }[];
               options: EmailPasswordAPIOptions;
               userContext: any;
-          }) => Promise<{
-              status: "OK";
-          }>);
+          }) => Promise<
+              | {
+                    status: "OK";
+                }
+              | GeneralErrorResponse
+          >);
 
     passwordResetPOST:
         | undefined
@@ -253,6 +257,7 @@ export type APIInterface = {
               | {
                     status: "RESET_PASSWORD_INVALID_TOKEN_ERROR";
                 }
+              | GeneralErrorResponse
           >);
 
     thirdPartySignInUpPOST:
@@ -273,10 +278,7 @@ export type APIInterface = {
                     session: SessionContainerInterface;
                     authCodeResponse: any;
                 }
-              | {
-                    status: "FIELD_ERROR";
-                    error: string;
-                }
+              | GeneralErrorResponse
               | {
                     status: "NO_EMAIL_GIVEN_BY_PROVIDER";
                 }
@@ -300,6 +302,7 @@ export type APIInterface = {
               | {
                     status: "WRONG_CREDENTIALS_ERROR";
                 }
+              | GeneralErrorResponse
           >);
 
     emailPasswordSignUpPOST:
@@ -320,6 +323,7 @@ export type APIInterface = {
               | {
                     status: "EMAIL_ALREADY_EXISTS_ERROR";
                 }
+              | GeneralErrorResponse
           >);
 
     appleRedirectHandlerPOST: