+++ lib/ts/recipe/thirdpartypasswordless/types.ts
@@ -38,6 +38,7 @@ import {
     TypeInput as SmsDeliveryTypeInput,
     TypeInputWithService as SmsDeliveryTypeInputWithService,
 } from "../../ingredients/smsdelivery/types";
+import { GeneralErrorResponse } from "../../types";
 
 export type DeviceType = DeviceTypeOriginal;
 
@@ -283,13 +284,7 @@ export type RecipeInterface = {
             isVerified: boolean;
         };
         userContext: any;
-    }): Promise<
-        | { status: "OK"; createdNewUser: boolean; user: User }
-        | {
-              status: "FIELD_ERROR";
-              error: string;
-          }
-    >;
+    }): Promise<{ status: "OK"; createdNewUser: boolean; user: User }>;
 
     createCode: (
         input: (
@@ -408,10 +403,13 @@ export type APIInterface = {
               provider: TypeProvider;
               options: ThirdPartyAPIOptions;
               userContext: any;
-          }) => Promise<{
-              status: "OK";
-              url: string;
-          }>);
+          }) => Promise<
+              | {
+                    status: "OK";
+                    url: string;
+                }
+              | GeneralErrorResponse
+          >);
 
     thirdPartySignInUpPOST:
         | undefined
@@ -431,10 +429,7 @@ export type APIInterface = {
                     session: SessionContainerInterface;
                     authCodeResponse: any;
                 }
-              | {
-                    status: "FIELD_ERROR";
-                    error: string;
-                }
+              | GeneralErrorResponse
               | {
                     status: "NO_EMAIL_GIVEN_BY_PROVIDER";
                 }
@@ -458,7 +453,7 @@ export type APIInterface = {
                     preAuthSessionId: string;
                     flowType: "USER_INPUT_CODE" | "MAGIC_LINK" | "USER_INPUT_CODE_AND_MAGIC_LINK";
                 }
-              | { status: "GENERAL_ERROR"; message: string }
+              | GeneralErrorResponse
           >);
 
     resendCodePOST:
@@ -468,7 +463,7 @@ export type APIInterface = {
                   options: PasswordlessAPIOptions;
                   userContext: any;
               }
-          ) => Promise<{ status: "GENERAL_ERROR"; message: string } | { status: "RESTART_FLOW_ERROR" | "OK" }>);
+          ) => Promise<GeneralErrorResponse | { status: "RESTART_FLOW_ERROR" | "OK" }>);
 
     consumeCodePOST:
         | undefined
@@ -499,7 +494,7 @@ export type APIInterface = {
                     failedCodeInputAttemptCount: number;
                     maximumCodeInputAttempts: number;
                 }
-              | { status: "GENERAL_ERROR"; message: string }
+              | GeneralErrorResponse
               | { status: "RESTART_FLOW_ERROR" }
           >);
 
@@ -509,10 +504,13 @@ export type APIInterface = {
               email: string;
               options: PasswordlessAPIOptions;
               userContext: any;
-          }) => Promise<{
-              status: "OK";
-              exists: boolean;
-          }>);
+          }) => Promise<
+              | {
+                    status: "OK";
+                    exists: boolean;
+                }
+              | GeneralErrorResponse
+          >);
 
     passwordlessUserPhoneNumberExistsGET:
         | undefined
@@ -520,10 +518,13 @@ export type APIInterface = {
               phoneNumber: string;
               options: PasswordlessAPIOptions;
               userContext: any;
-          }) => Promise<{
-              status: "OK";
-              exists: boolean;
-          }>);
+          }) => Promise<
+              | {
+                    status: "OK";
+                    exists: boolean;
+                }
+              | GeneralErrorResponse
+          >);
 };
 
 export type TypeThirdPartyPasswordlessEmailDeliveryInput =